Split Operations into Operations + FP
